GM-15 소셜 로그인 저장 및 리팩토링 진행
