+++ build.gradle
@@ -29,7 +29,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-security'
 
 //    implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
-
+    implementation("com.github.gavlyukovskiy:p6spy-spring-boot-starter:1.8.0")
     implementation 'io.jsonwebtoken:jjwt:0.9.1'
